[FEAT] image를 deatil하게 보여주는 viewmodel의 로직 통일(#39)
