+++ Genti_iOS/Genti_iOS/Complete/Component/VerticalImageContentView.swift
@@ -47,29 +47,29 @@ struct VerticalImageContentView: View {
                     Spacer()
                         .frame(height: 16)
                     
-                    Image("testImage")
-                        .resizable()
-                        .aspectRatio(contentMode: .fit)
-                        .frame(height: 360)
-                        .overlay(alignment: .bottomTrailing) {
-                            Image("Download")
-                                .resizable()
-                                .frame(width: 44, height: 44)
-                                .padding(.bottom, 12)
-                                .padding(.trailing, 22)
-                                .onTapGesture {
-                                    print(#fileID, #function, #line, "- downloadtap")
-                                }
-                        }
-                        .clipShape(RoundedRectangle(cornerRadius: 15))
-                        .overlay(
-                            RoundedRectangle(cornerRadius: 15)
-                                .strokeBorder(LinearGradient(colors: [.gentiGreen, .gentiGreen.opacity(0)], startPoint: .topLeading, endPoint: .bottomTrailing), lineWidth: 2)
-                        )
-                        .onTapGesture {
-                            print(#fileID, #function, #line, "- imagetap")
-                            self.viewModel.sendAction(.imageTap)
-                        }
+                viewModel.getImage
+                    .resizable()
+                    .aspectRatio(contentMode: .fit)
+                    .frame(height: 360)
+                    .overlay(alignment: .bottomTrailing) {
+                        Image("Download")
+                            .resizable()
+                            .frame(width: 44, height: 44)
+                            .padding(.bottom, 12)
+                            .padding(.trailing, 22)
+                            .onTapGesture {
+                                print(#fileID, #function, #line, "- downloadtap")
+                            }
+                    }
+                    .clipShape(RoundedRectangle(cornerRadius: 15))
+                    .overlay(
+                        RoundedRectangle(cornerRadius: 15)
+                            .strokeBorder(LinearGradient(colors: [.gentiGreen, .gentiGreen.opacity(0)], startPoint: .topLeading, endPoint: .bottomTrailing), lineWidth: 2)
+                    )
+                    .onTapGesture {
+                        print(#fileID, #function, #line, "- imagetap")
+                        self.viewModel.sendAction(.imageTap)
+                    }
                     
                     Spacer()
                         .frame(height: 18)
@@ -79,9 +79,12 @@ struct VerticalImageContentView: View {
                 }
         } //:ZSTACK
         .frame(height: 640)
+        .onAppear {
+            self.viewModel.sendAction(.viewWillAppear)
+        }
     }
 }
 
-#Preview {
-    VerticalImageContentView(viewModel: PhotoCompleteViewViewModel(router: .init()))
-}
+//#Preview {
+//    VerticalImageContentView(viewModel: PhotoCompleteViewViewModel(router: .init()))
+//}
