+++ Genti_iOS/Genti_iOS/Complete/Model/CompletePhotoEntity.swift
@@ -1,0 +1,13 @@
+//
+//  CompletePhotoEntity.swift
+//  Genti_iOS
+//
+//  Created by uiskim on 7/7/24.
+//
+
+import Foundation
+
+struct CompletePhotoEntity {
+    var imageUrlString: String = "https://d2rvmd5lmgmzuf.cloudfront.net/ADMIN_UPLOADED_IMAGE/어드민응답1.png-f8614bc9-3614-40bb-a91e-f8398be66d7c"
+    var imageRatio: PhotoRatio = .twoThird
+}
